+++ src/parser/monster/source.js
@@ -1,3 +1,4 @@
+import { DICTIONARY } from "../../config/_module.mjs";
 import { DDBSources } from "../../lib/_module.mjs";
 import DDBMonster from "../DDBMonster.js";
 
@@ -15,4 +16,12 @@ DDBMonster.prototype._generateSource = function _generateSource() {
   DDBSources.tweakSourceData(source);
 
   this.npc.system.details.source = source;
+
+  this.legacy = CONFIG.DDB.sources.some((ddbSource) =>
+    DICTIONARY.sourceCategories.legacy.includes(ddbSource.sourceCategoryId),
+  );
+  this.is2014 = Number.isInteger(ddbSource.id) && ddbSource.id < 145;
+  this.is2024 = !this.is2014;
+
+  this.npc.system.source.rules = this.is2014 ? "2014" : "2024";
 };